+++ exam/src/main/java/com/projectexam/exam/Controllers/PatientController.java
@@ -9,7 +9,7 @@ import com.projectexam.exam.Services.PatientService;
 
 @RestController
 @RequestMapping("/patient")
-public class PatientController extends GenericController<PatientDto, Integer, PatientService> {
+public class PatientController extends GenericController<PatientDto, Long, PatientService> {
     public PatientController(PatientService service) {
         super(service);
     }
