+++ exam/src/main/java/com/projectexam/exam/Services/MedecinService.java
@@ -3,6 +3,6 @@ package com.projectexam.exam.Services;
 import com.projectexam.exam.Dtos.MedecinDto;
 import com.projectexam.exam.Generic.GenericService;
 
-public interface MedecinService extends GenericService<MedecinDto, Integer> {
+public interface MedecinService extends GenericService<MedecinDto, Long> {
     
 }
