+++ exam/src/main/java/com/projectexam/exam/Services/MedicamentService.java
@@ -3,6 +3,6 @@ package com.projectexam.exam.Services;
 import com.projectexam.exam.Dtos.MedicamentDto;
 import com.projectexam.exam.Generic.GenericService;
 
-public interface MedicamentService extends GenericService<MedicamentDto, Integer> {
+public interface MedicamentService extends GenericService<MedicamentDto, Long> {
     
 }
