+++ exam/src/main/java/com/projectexam/exam/Services/PatientService.java
@@ -3,6 +3,6 @@ package com.projectexam.exam.Services;
 import com.projectexam.exam.Dtos.PatientDto;
 import com.projectexam.exam.Generic.GenericService;
 
-public interface PatientService extends GenericService<PatientDto, Integer> {
+public interface PatientService extends GenericService<PatientDto, Long> {
     
 }
